+++ codegens/curl/test/unit/convert.test.js
@@ -1,5 +1,7 @@
-var expect = require('chai').expect,
-  sdk = require('postman-collection'),
+var _ = require('lodash'),
+  expect = require('chai').expect,
+  { Request } = require('postman-collection/lib/collection/request'),
+  { Url } = require('postman-collection/lib/collection/url'),
   convert = require('../../index').convert,
   getUrlStringfromUrlObject = require('../../lib/util').getUrlStringfromUrlObject;
 
@@ -9,7 +11,7 @@ describe('curl convert function', function () {
 
     it('should return snippet with carat(^) as line continuation ' +
             'character for multiline code generation', function () {
-      request = new sdk.Request({
+      request = new Request({
         'method': 'POST',
         'header': [],
         'body': {
@@ -35,7 +37,7 @@ describe('curl convert function', function () {
     });
 
     it('should return snippet with url in single quote(\')', function () {
-      request = new sdk.Request({
+      request = new Request({
         'method': 'POST',
         'header': [],
         'body': {
@@ -57,7 +59,7 @@ describe('curl convert function', function () {
     });
 
     it('should return snippet with url in double quote(")', function () {
-      request = new sdk.Request({
+      request = new Request({
         'method': 'POST',
         'header': [],
         'body': {
@@ -79,7 +81,7 @@ describe('curl convert function', function () {
     });
 
     it('should add semicolon after header key, if the value is empty string', function () {
-      request = new sdk.Request({
+      request = new Request({
         'method': 'GET',
         'header': [
           {
@@ -110,7 +112,7 @@ describe('curl convert function', function () {
 
     it('should return snippet with backslash(\\) as line continuation ' +
             'character for multiline code generation by default', function () {
-      request = new sdk.Request({
+      request = new Request({
         'method': 'POST',
         'header': [],
         'body': {
@@ -136,7 +138,7 @@ describe('curl convert function', function () {
 
     it('should return snippet with backtick(`) as line continuation ' +
             'character for multiline code generation', function () {
-      request = new sdk.Request({
+      request = new Request({
         'method': 'POST',
         'header': [],
         'body': {
@@ -162,7 +164,7 @@ describe('curl convert function', function () {
     });
 
     it('should add content type if formdata field contains a content-type', function () {
-      request = new sdk.Request({
+      request = new Request({
         'method': 'POST',
         'body': {
           'mode': 'formdata',
@@ -198,7 +200,7 @@ describe('curl convert function', function () {
     });
 
     it('should parse header with string value properly', function () {
-      request = new sdk.Request({
+      request = new Request({
         'method': 'POST',
         'header': [
           {
@@ -230,7 +232,7 @@ describe('curl convert function', function () {
         ']world',
         'world}'
       ].forEach(function (value) {
-        const request = new sdk.Request({
+        const request = new Request({
           'method': 'GET',
           'url': {
             'raw': `http://example.com?hello=${value}`,
@@ -263,7 +265,7 @@ describe('curl convert function', function () {
     });
 
     it('should return snippet without errors when request object has no body property', function () {
-      request = new sdk.Request({
+      request = new Request({
         'method': 'GET',
         'header': [],
         'url': {
@@ -289,7 +291,7 @@ describe('curl convert function', function () {
 
     it('should return snippet with backslash(\\) character as line continuation ' +
          'character for multiline code generation', function () {
-      request = new sdk.Request({
+      request = new Request({
         'method': 'POST',
         'header': [],
         'body': {
@@ -316,7 +318,7 @@ describe('curl convert function', function () {
 
     it('should not encode queryParam unresolved variables and ' +
     'leave it inside double parenthesis {{xyz}}', function () {
-      request = new sdk.Request({
+      request = new Request({
         'method': 'POST',
         'header': [],
         'url': {
@@ -349,7 +351,7 @@ describe('curl convert function', function () {
     });
 
     it('should encode queryParams other than unresolved variables', function () {
-      request = new sdk.Request({
+      request = new Request({
         'method': 'POST',
         'header': [],
         'url': {
@@ -382,7 +384,7 @@ describe('curl convert function', function () {
     });
 
     it('should trim header keys and not trim header values', function () {
-      var request = new sdk.Request({
+      var request = new Request({
         'method': 'GET',
         'header': [
           {
@@ -411,7 +413,7 @@ describe('curl convert function', function () {
     });
 
     it('should generate snippets for no files in form data', function () {
-      var request = new sdk.Request({
+      var request = new Request({
         'method': 'POST',
         'header': [],
         'body': {
@@ -461,7 +463,7 @@ describe('curl convert function', function () {
 
     it('should generate valid snippets for single/double quotes in URL', function () {
       // url = https://a"b'c.com/'d/"e
-      var request = new sdk.Request("https://a\"b'c.com/'d/\"e"); // eslint-disable-line quotes
+      var request = new Request("https://a\"b'c.com/'d/\"e"); // eslint-disable-line quotes
       convert(request, {}, function (error, snippet) {
         if (error) {
           expect.fail(null, null, error);
@@ -474,7 +476,7 @@ describe('curl convert function', function () {
 
     it('should generate valid snippets when quoteType is "double"', function () {
       // url = https://a"b'c.com/'d/"e
-      var request = new sdk.Request({
+      var request = new Request({
         'method': 'POST',
         'body': {
           'mode': 'formdata',
@@ -506,7 +508,7 @@ describe('curl convert function', function () {
     });
 
     it('should not add appropriate escaping characters when quote type is "double"', function () {
-      var request = new sdk.Request({
+      var request = new Request({
         'method': 'POST',
         'header': [],
         'body': {
@@ -538,7 +540,7 @@ describe('curl convert function', function () {
     });
 
     it('should escape special characters when quoteType is "double"', function () {
-      var request = new sdk.Request({
+      var request = new Request({
         'method': 'POST',
         'header': [],
         'body': {
@@ -572,7 +574,7 @@ describe('curl convert function', function () {
     });
 
     it('should longer option for body even if longFormat is disabled if @ character is present', function () {
-      let request = new sdk.Request({
+      let request = new Request({
         'method': 'POST',
         'header': [],
         'body': {
@@ -606,7 +608,7 @@ describe('curl convert function', function () {
 
       it('should return empty string for an url object for an empty url or if no url object is passed', function () {
         rawUrl = '';
-        urlObject = new sdk.Url(rawUrl);
+        urlObject = new Url(rawUrl);
         outputUrlString = getUrlStringfromUrlObject(urlObject);
         expect(outputUrlString).to.be.empty;
         outputUrlString = getUrlStringfromUrlObject();
@@ -615,14 +617,14 @@ describe('curl convert function', function () {
 
       it('should add protocol if present in the url object', function () {
         rawUrl = 'https://postman-echo.com';
-        urlObject = new sdk.Url(rawUrl);
+        urlObject = new Url(rawUrl);
         outputUrlString = getUrlStringfromUrlObject(urlObject);
         expect(outputUrlString).to.equal(rawUrl);
       });
 
       it('should add the auth information if present in the url object', function () {
         rawUrl = 'https://user:password@postman-echo.com';
-        urlObject = new sdk.Url(rawUrl);
+        urlObject = new Url(rawUrl);
         outputUrlString = getUrlStringfromUrlObject(urlObject);
         expect(outputUrlString).to.equal(rawUrl);
       });
@@ -630,28 +632,28 @@ describe('curl convert function', function () {
       it('should not add the auth information if user isn\'t present but' +
       ' password is present in the url object', function () {
         rawUrl = 'https://:password@postman-echo.com';
-        urlObject = new sdk.Url(rawUrl);
+        urlObject = new Url(rawUrl);
         outputUrlString = getUrlStringfromUrlObject(urlObject);
         expect(outputUrlString).to.not.include(':password');
       });
 
       it('should add host if present in the url object', function () {
         rawUrl = 'https://postman-echo.com';
-        urlObject = new sdk.Url(rawUrl);
+        urlObject = new Url(rawUrl);
         outputUrlString = getUrlStringfromUrlObject(urlObject);
         expect(outputUrlString).to.equal(rawUrl);
       });
 
       it('should add port if present in the url object', function () {
         rawUrl = 'https://postman-echo.com:8080';
-        urlObject = new sdk.Url(rawUrl);
+        urlObject = new Url(rawUrl);
         outputUrlString = getUrlStringfromUrlObject(urlObject);
         expect(outputUrlString).to.equal(rawUrl);
       });
 
       it('should add path if present in the url object', function () {
         rawUrl = 'https://postman-echo.com/get';
-        urlObject = new sdk.Url(rawUrl);
+        urlObject = new Url(rawUrl);
         outputUrlString = getUrlStringfromUrlObject(urlObject);
         expect(outputUrlString).to.equal(rawUrl);
       });
@@ -660,7 +662,7 @@ describe('curl convert function', function () {
 
         it('should not encode unresolved query params', function () {
           rawUrl = 'https://postman-echo.com/get?key={{value}}';
-          urlObject = new sdk.Url(rawUrl);
+          urlObject = new Url(rawUrl);
           outputUrlString = getUrlStringfromUrlObject(urlObject);
           expect(outputUrlString).to.not.include('key=%7B%7Bvalue%7B%7B');
           expect(outputUrlString).to.equal(rawUrl);
@@ -668,7 +670,7 @@ describe('curl convert function', function () {
 
         it('should encode query params other than unresolved variables', function () {
           rawUrl = 'https://postman-echo.com/get?key=\'a b c\'';
-          urlObject = new sdk.Url(rawUrl);
+          urlObject = new Url(rawUrl);
           outputUrlString = getUrlStringfromUrlObject(urlObject);
           expect(outputUrlString).to.not.include('key=\'a b c\'');
           expect(outputUrlString).to.equal('https://postman-echo.com/get?key=%27a%20b%20c%27');
@@ -677,7 +679,7 @@ describe('curl convert function', function () {
         it('should not encode unresolved query params and ' +
         'encode every other query param, both present together', function () {
           rawUrl = 'https://postman-echo.com/get?key1={{value}}&key2=\'a b+c\'';
-          urlObject = new sdk.Url(rawUrl);
+          urlObject = new Url(rawUrl);
           outputUrlString = getUrlStringfromUrlObject(urlObject);
           expect(outputUrlString).to.not.include('key1=%7B%7Bvalue%7B%7B');
           expect(outputUrlString).to.not.include('key2=\'a b+c\'');
@@ -686,13 +688,13 @@ describe('curl convert function', function () {
 
         it('should not encode query params that are already encoded', function () {
           rawUrl = 'https://postman-echo.com/get?query=urn%3Ali%3Afoo%3A62324';
-          urlObject = new sdk.Url(rawUrl);
+          urlObject = new Url(rawUrl);
           outputUrlString = getUrlStringfromUrlObject(urlObject);
           expect(outputUrlString).to.equal('https://postman-echo.com/get?query=urn%3Ali%3Afoo%3A62324');
         });
 
         it('should discard disabled query params', function () {
-          urlObject = new sdk.Url({
+          urlObject = new Url({
             protocol: 'https',
             host: 'postman-echo.com',
             query: [
@@ -707,14 +709,14 @@ describe('curl convert function', function () {
 
       it('should add hash if present in the url object', function () {
         rawUrl = 'https://postmanm-echo.com/get#hash';
-        urlObject = new sdk.Url(rawUrl);
+        urlObject = new Url(rawUrl);
         outputUrlString = getUrlStringfromUrlObject(urlObject);
         expect(outputUrlString).to.equal(rawUrl);
       });
     });
 
     it('should not add --request parameter in POST request if body is present', function () {
-      var request = new sdk.Request({
+      var request = new Request({
         'method': 'POST',
         'header': [],
         'body': {
@@ -747,7 +749,7 @@ describe('curl convert function', function () {
     });
 
     it('should add --request parameter in POST request if body is not present', function () {
-      var request = new sdk.Request({
+      var request = new Request({
         'method': 'POST',
         'header': [],
         'url': {
@@ -773,7 +775,7 @@ describe('curl convert function', function () {
     });
 
     it('should add --request parameter in GET request if body is present', function () {
-      var request = new sdk.Request({
+      var request = new Request({
         'method': 'GET',
         'header': [],
         'body': {
@@ -807,7 +809,7 @@ describe('curl convert function', function () {
 
     it('should not add --request parameter in GET request if body is present ' +
       'but disableBodyPruning is false', function () {
-      const request = new sdk.Request({
+      const request = new Request({
         'method': 'GET',
         'header': [],
         'body': {
@@ -846,7 +848,7 @@ describe('curl convert function', function () {
 
     describe('followRedirect and followOriginalHttpMethod', function () {
       it('should add --request parameter when passed true via options', function () {
-        const request = new sdk.Request({
+        const request = new Request({
           'method': 'POST',
           'header': [],
           'body': {
@@ -879,7 +881,7 @@ describe('curl convert function', function () {
       });
 
       it('should not add --request parameter when passed false via options', function () {
-        const request = new sdk.Request({
+        const request = new Request({
           'method': 'POST',
           'header': [],
           'body': {
@@ -912,7 +914,7 @@ describe('curl convert function', function () {
       });
 
       it('should add --request parameter when passed false via options but true in request settings', function () {
-        const request = new sdk.Request({
+        const request = new Request({
           'method': 'POST',
           'header': [],
           'body': {
@@ -951,7 +953,7 @@ describe('curl convert function', function () {
       });
 
       it('should not add --request parameter when passed true via options but false in request settings', function () {
-        const request = new sdk.Request({
+        const request = new Request({
           'method': 'POST',
           'header': [],
           'body': {
@@ -990,7 +992,7 @@ describe('curl convert function', function () {
       });
 
       it('should work when protocolProfileBehavior is null in request settings', function () {
-        const request = new sdk.Request({
+        const request = new Request({
           'method': 'POST',
           'header': [],
           'body': {
@@ -1022,6 +1024,123 @@ describe('curl convert function', function () {
           }
           expect(snippet).to.be.a('string');
           expect(snippet).to.include('--request POST');
+        });
+      });
+    });
+
+    describe('should correctly handle NTLM auth', function () {
+      const sampleRequest = {
+        'method': 'POST',
+        'header': [],
+        'auth': {
+          'type': 'ntlm',
+          'ntlm': []
+        },
+        'url': {
+          'raw': 'https://postman-echo.com/post',
+          'protocol': 'https',
+          'host': [
+            'postman-echo',
+            'com'
+          ],
+          'path': [
+            'post'
+          ]
+        }
+      };
+
+      it('when no username or password is present', function () {
+        const request = new Request(sampleRequest);
+
+        convert(request, {}, function (error, snippet) {
+          if (error) {
+            expect.fail(null, null, error);
+          }
+          expect(snippet).to.be.a('string');
+          expect(snippet).to.not.include('--ntlm');
+        });
+      });
+
+      it('when empty username and password is present', function () {
+        const request = new Request(Object.assign({ auth: {
+          'type': 'ntlm',
+          'ntlm': [
+            {key: 'username', value: ''},
+            {key: 'password', value: ''}
+          ]
+        }}, sampleRequest));
+
+        convert(request, {}, function (error, snippet) {
+          if (error) {
+            expect.fail(null, null, error);
+          }
+          expect(snippet).to.be.a('string');
+          expect(snippet).to.not.include('--ntlm');
+        });
+      });
+
+      it('when correct username and password is present with single quotes as option', function () {
+        const request = new Request(_.set(sampleRequest, 'auth.ntlm', [
+          {key: 'username', value: 'joh\'n'},
+          {key: 'password', value: 'tennesse"e'}
+        ]));
+
+        convert(request, { quoteType: 'single' }, function (error, snippet) {
+          if (error) {
+            expect.fail(null, null, error);
+          }
+          expect(snippet).to.be.a('string');
+          expect(snippet).to.equal('curl --ntlm --user \'joh\'\\\'\'n:tennesse"e\' --location' +
+            ' --request POST \'https://postman-echo.com/post\'');
+        });
+      });
+
+      it('when correct username and password is present with double as option', function () {
+        const request = new Request(_.set(sampleRequest, 'auth.ntlm', [
+          {key: 'username', value: 'joh\'n'},
+          {key: 'password', value: 'tennesse"e'}
+        ]));
+
+        convert(request, { quoteType: 'double' }, function (error, snippet) {
+          if (error) {
+            expect.fail(null, null, error);
+          }
+          expect(snippet).to.be.a('string');
+          expect(snippet).to.equal('curl --ntlm --user "joh\'n:tennesse\\"e" --location' +
+            ' --request POST "https://postman-echo.com/post"');
+        });
+      });
+
+      it('when correct username and password is present with long format option disabled', function () {
+        const request = new Request(_.set(sampleRequest, 'auth.ntlm', [
+          {key: 'username', value: 'joh\'n'},
+          {key: 'password', value: 'tennesse"e'}
+        ]));
+
+        convert(request, { longFormat: false }, function (error, snippet) {
+          if (error) {
+            expect.fail(null, null, error);
+          }
+          expect(snippet).to.be.a('string');
+          expect(snippet).to.equal('curl --ntlm -u \'joh\'\\\'\'n:tennesse"e\' -L' +
+            ' -X POST \'https://postman-echo.com/post\'');
+        });
+      });
+
+      it('when username and password is present with domain as well', function () {
+        const request = new Request(_.set(sampleRequest, 'auth.ntlm', [
+          {key: 'username', value: 'joh\'n'},
+          {key: 'password', value: 'tennesse"e'},
+          {key: 'domain', value: 'radio'}
+        ]));
+
+        convert(request, {}, function (error, snippet) {
+          if (error) {
+            expect.fail(null, null, error);
+          }
+          expect(snippet).to.be.a('string');
+          expect(snippet).to.equal('curl --ntlm --user \'radio\\joh\'\\\'\'n:tennesse"e\' --location' +
+            ' --request POST \'https://postman-echo.com/post\'');
         });
       });
     });