+++ codegens/java-unirest/test/unit/convert.test.js
@@ -1,5 +1,6 @@
 var expect = require('chai').expect,
-  sdk = require('postman-collection'),
+  { Request } = require('postman-collection/lib/collection/request'),
+  { Url } = require('postman-collection/lib/collection/url'),
   convert = require('../../lib/index').convert,
   sanitize = require('../../lib/util').sanitize,
   getUrlStringfromUrlObject = require('../../lib/parseRequest').getUrlStringfromUrlObject,
@@ -18,7 +19,7 @@ describe('java unirest convert function for test collection', function () {
       line_no;
 
     it('should return a Tab indented snippet ', function () {
-      request = new sdk.Request(mainCollection.item[0].request);
+      request = new Request(mainCollection.item[0].request);
       options = {
         indentType: 'Tab',
         indentCount: 1
@@ -43,7 +44,7 @@ describe('java unirest convert function for test collection', function () {
     });
 
     it('should return snippet with setTimeouts function when timeout is set to non zero', function () {
-      request = new sdk.Request(mainCollection.item[0].request);
+      request = new Request(mainCollection.item[0].request);
       options = {
         requestTimeout: 1000
       };
@@ -59,7 +60,7 @@ describe('java unirest convert function for test collection', function () {
 
     it('should return snippet with setTimeouts function setting both ' +
             'connection and socket timeout to 0 when requestTimeout is set to 0', function () {
-      request = new sdk.Request(mainCollection.item[0].request);
+      request = new Request(mainCollection.item[0].request);
       options = {
         requestTimeout: 0
       };
@@ -75,7 +76,7 @@ describe('java unirest convert function for test collection', function () {
 
     it('should return snippet with disableRedirectHandling function for' +
             'follow redirect option set to false', function () {
-      request = new sdk.Request(mainCollection.item[0].request);
+      request = new Request(mainCollection.item[0].request);
       options = {
         followRedirect: false
       };
@@ -90,7 +91,7 @@ describe('java unirest convert function for test collection', function () {
     });
 
     it('should add content type if formdata field contains a content-type', function () {
-      request = new sdk.Request({
+      request = new Request({
         'method': 'POST',
         'body': {
           'mode': 'formdata',
@@ -126,7 +127,7 @@ describe('java unirest convert function for test collection', function () {
 
     it('should include import statements, main class and print statements ' +
             'when includeBoilerplate is set to true', function () {
-      request = new sdk.Request(mainCollection.item[0].request);
+      request = new Request(mainCollection.item[0].request);
       options = {
         includeBoilerplate: true,
         indentType: 'Tab',
@@ -156,7 +157,7 @@ describe('java unirest convert function for test collection', function () {
         'url': 'https://echo.getpostman.com/post',
         'method': 'POST'
       };
-      request = new sdk.Request(reqObject);
+      request = new Request(reqObject);
       options = {};
       convert(request, options, function (error, snippet) {
         if (error) {
@@ -177,7 +178,7 @@ describe('java unirest convert function for test collection', function () {
         'url': 'https://postman-echo.com/request',
         'method': 'PROPFIND'
       };
-      request = new sdk.Request(reqObject);
+      request = new Request(reqObject);
       options = {};
       convert(request, options, function (error, snippet) {
         if (error) {
@@ -192,7 +193,7 @@ describe('java unirest convert function for test collection', function () {
 
     it('should not encode queryParam unresolved variables and ' +
     'leave it inside double parenthesis {{xyz}}', function () {
-      request = new sdk.Request({
+      request = new Request({
         'method': 'POST',
         'header': [],
         'url': {
@@ -224,7 +225,7 @@ describe('java unirest convert function for test collection', function () {
     });
 
     it('should encode queryParams other than unresolved variables', function () {
-      request = new sdk.Request({
+      request = new Request({
         'method': 'POST',
         'header': [],
         'url': {
@@ -256,7 +257,7 @@ describe('java unirest convert function for test collection', function () {
     });
 
     it('should trim header keys and not trim header values', function () {
-      var request = new sdk.Request({
+      var request = new Request({
         'method': 'GET',
         'header': [
           {
@@ -284,7 +285,7 @@ describe('java unirest convert function for test collection', function () {
 
     it('should add the force multipart body method when ' +
       'there are no files but contains text field params in multipart/form-data', function () {
-      var request = new sdk.Request({
+      var request = new Request({
         'method': 'GET',
         'body': {
           'mode': 'formdata',
@@ -323,7 +324,7 @@ describe('java unirest convert function for test collection', function () {
 
     it('should not add the force multipart body method when ' +
       'there are file fields in multipart/form-data', function () {
-      var request = new sdk.Request({
+      var request = new Request({
         'method': 'GET',
         'body': {
           'mode': 'formdata',
@@ -359,7 +360,7 @@ describe('java unirest convert function for test collection', function () {
     });
 
     it('should generate snippets for no files in form data', function () {
-      var request = new sdk.Request({
+      var request = new Request({
         'method': 'POST',
         'header': [],
         'body': {
@@ -407,7 +408,7 @@ describe('java unirest convert function for test collection', function () {
 
     it('should generate valid snippets for single/double quotes in URL', function () {
       // url = https://a"b'c.com/'d/"e
-      var request = new sdk.Request("https://a\"b'c.com/'d/\"e"); // eslint-disable-line quotes
+      var request = new Request("https://a\"b'c.com/'d/\"e"); // eslint-disable-line quotes
       convert(request, {}, function (error, snippet) {
         if (error) {
           expect.fail(null, null, error);
@@ -423,7 +424,7 @@ describe('java unirest convert function for test collection', function () {
 
     it('should return empty string for an url object for an empty url or if no url object is passed', function () {
       rawUrl = '';
-      urlObject = new sdk.Url(rawUrl);
+      urlObject = new Url(rawUrl);
       outputUrlString = getUrlStringfromUrlObject(urlObject);
       expect(outputUrlString).to.be.empty;
       outputUrlString = getUrlStringfromUrlObject();
@@ -432,14 +433,14 @@ describe('java unirest convert function for test collection', function () {
 
     it('should add protocol if present in the url object', function () {
       rawUrl = 'https://postman-echo.com';
-      urlObject = new sdk.Url(rawUrl);
+      urlObject = new Url(rawUrl);
       outputUrlString = getUrlStringfromUrlObject(urlObject);
       expect(outputUrlString).to.equal(rawUrl);
     });
 
     it('should add the auth information if present in the url object', function () {
       rawUrl = 'https://user:password@postman-echo.com';
-      urlObject = new sdk.Url(rawUrl);
+      urlObject = new Url(rawUrl);
       outputUrlString = getUrlStringfromUrlObject(urlObject);
       expect(outputUrlString).to.equal(rawUrl);
     });
@@ -447,28 +448,28 @@ describe('java unirest convert function for test collection', function () {
     it('should not add the auth information if user isn\'t present but' +
     ' password is present in the url object', function () {
       rawUrl = 'https://:password@postman-echo.com';
-      urlObject = new sdk.Url(rawUrl);
+      urlObject = new Url(rawUrl);
       outputUrlString = getUrlStringfromUrlObject(urlObject);
       expect(outputUrlString).to.not.include(':password');
     });
 
     it('should add host if present in the url object', function () {
       rawUrl = 'https://postman-echo.com';
-      urlObject = new sdk.Url(rawUrl);
+      urlObject = new Url(rawUrl);
       outputUrlString = getUrlStringfromUrlObject(urlObject);
       expect(outputUrlString).to.equal(rawUrl);
     });
 
     it('should add port if present in the url object', function () {
       rawUrl = 'https://postman-echo.com:8080';
-      urlObject = new sdk.Url(rawUrl);
+      urlObject = new Url(rawUrl);
       outputUrlString = getUrlStringfromUrlObject(urlObject);
       expect(outputUrlString).to.equal(rawUrl);
     });
 
     it('should add path if present in the url object', function () {
       rawUrl = 'https://postman-echo.com/get';
-      urlObject = new sdk.Url(rawUrl);
+      urlObject = new Url(rawUrl);
       outputUrlString = getUrlStringfromUrlObject(urlObject);
       expect(outputUrlString).to.equal(rawUrl);
     });
@@ -477,7 +478,7 @@ describe('java unirest convert function for test collection', function () {
 
       it('should not encode unresolved query params', function () {
         rawUrl = 'https://postman-echo.com/get?key={{value}}';
-        urlObject = new sdk.Url(rawUrl);
+        urlObject = new Url(rawUrl);
         outputUrlString = getUrlStringfromUrlObject(urlObject);
         expect(outputUrlString).to.include('key=%7B%7Bvalue%7D%7D');
         expect(outputUrlString).to.equal('https://postman-echo.com/get?key=%7B%7Bvalue%7D%7D');
@@ -485,7 +486,7 @@ describe('java unirest convert function for test collection', function () {
 
       it('should encode query params other than unresolved variables', function () {
         rawUrl = 'https://postman-echo.com/get?key=\'a b c\'';
-        urlObject = new sdk.Url(rawUrl);
+        urlObject = new Url(rawUrl);
         outputUrlString = getUrlStringfromUrlObject(urlObject);
         expect(outputUrlString).to.not.include('key=\'a b c\'');
         expect(outputUrlString).to.equal('https://postman-echo.com/get?key=%27a%20b%20c%27');
@@ -493,7 +494,7 @@ describe('java unirest convert function for test collection', function () {
 
       it('should not encode query params that are already encoded', function () {
         rawUrl = 'https://postman-echo.com/get?query=urn%3Ali%3Afoo%3A62324';
-        urlObject = new sdk.Url(rawUrl);
+        urlObject = new Url(rawUrl);
         outputUrlString = getUrlStringfromUrlObject(urlObject);
         expect(outputUrlString).to.equal('https://postman-echo.com/get?query=urn%3Ali%3Afoo%3A62324');
       });
@@ -501,7 +502,7 @@ describe('java unirest convert function for test collection', function () {
       it('should not encode unresolved query params and ' +
       'encode every other query param, both present together', function () {
         rawUrl = 'https://postman-echo.com/get?key1={{value}}&key2=\'a b+c\'';
-        urlObject = new sdk.Url(rawUrl);
+        urlObject = new Url(rawUrl);
         outputUrlString = getUrlStringfromUrlObject(urlObject);
         expect(outputUrlString).to.include('key1=%7B%7Bvalue%7D%7D');
         expect(outputUrlString).to.not.include('key2=\'a b+c\'');
@@ -509,7 +510,7 @@ describe('java unirest convert function for test collection', function () {
       });
 
       it('should discard disabled query params', function () {
-        urlObject = new sdk.Url({
+        urlObject = new Url({
           protocol: 'https',
           host: 'postman-echo.com',
           query: [
@@ -524,7 +525,7 @@ describe('java unirest convert function for test collection', function () {
 
     it('should add hash if present in the url object', function () {
       rawUrl = 'https://postmanm-echo.com/get#hash';
-      urlObject = new sdk.Url(rawUrl);
+      urlObject = new Url(rawUrl);
       outputUrlString = getUrlStringfromUrlObject(urlObject);
       expect(outputUrlString).to.equal(rawUrl);
     });