+++ codegens/php-curl/test/unit/converter.test.js
@@ -1,5 +1,6 @@
 var expect = require('chai').expect,
-  sdk = require('postman-collection'),
+  { Request } = require('postman-collection/lib/collection/request'),
+  { Url } = require('postman-collection/lib/collection/url'),
   convert = require('../../lib/index').convert,
   getUrlStringfromUrlObject = require('../../lib/util/sanitize').getUrlStringfromUrlObject;
 
@@ -10,7 +11,7 @@ describe('php-curl converter', function () {
   });
 
   it('should trim header keys and not trim header values', function () {
-    var request = new sdk.Request({
+    var request = new Request({
       'method': 'GET',
       'header': [
         {
@@ -38,7 +39,7 @@ describe('php-curl converter', function () {
   });
 
   it('should generate snippets for no files in form data', function () {
-    var request = new sdk.Request({
+    var request = new Request({
       'method': 'POST',
       'header': [],
       'body': {
@@ -89,7 +90,7 @@ describe('php-curl converter', function () {
   it('should not encode unresolved query params and ' +
   'encode every other query param, both present together', function () {
     let rawUrl = 'https://postman-echo.com/get?key1={{value}}&key2=\'a b+c\'',
-      urlObject = new sdk.Url(rawUrl),
+      urlObject = new Url(rawUrl),
       outputUrlString = getUrlStringfromUrlObject(urlObject);
     expect(outputUrlString).to.not.include('key1=%7B%7Bvalue%7B%7B');
     expect(outputUrlString).to.not.include('key2=\'a b+c\'');