+++ codegens/php-guzzle/test/unit/parseBody.test.js
@@ -1,5 +1,5 @@
 var expect = require('chai').expect,
-  sdk = require('postman-collection'),
+  { Collection } = require('postman-collection/lib/collection/collection'),
   fs = require('fs'),
   path = require('path'),
   {
@@ -10,7 +10,7 @@ var expect = require('chai').expect,
 describe('parseBody function', function () {
 
   it('should parse a raw json Body', function () {
-    const collection = new sdk.Collection(JSON.parse(
+    const collection = new Collection(JSON.parse(
         fs.readFileSync(path.resolve(__dirname, collectionsPath, './sample_collection.json').toString()))),
       body = collection.items.members[8].request.body,
       indentation = '  ',
@@ -21,7 +21,7 @@ describe('parseBody function', function () {
   });
 
   it('should parse a raw json Body with indentation', function () {
-    const collection = new sdk.Collection(JSON.parse(
+    const collection = new Collection(JSON.parse(
         fs.readFileSync(path.resolve(__dirname, collectionsPath, './sample_collection.json').toString()))),
       body = collection.items.members[8].request.body,
       indentation = '   ',
@@ -32,7 +32,7 @@ describe('parseBody function', function () {
   });
 
   it('should parse a raw xml Body', function () {
-    const collection = new sdk.Collection(JSON.parse(
+    const collection = new Collection(JSON.parse(
         fs.readFileSync(path.resolve(__dirname, collectionsPath, './sample_collection.json').toString()))),
       body = collection.items.members[10].request.body,
       indentation = '   ',
@@ -43,7 +43,7 @@ describe('parseBody function', function () {
   });
 
   it('should return form-url-encoded params', function () {
-    const collection = new sdk.Collection(JSON.parse(
+    const collection = new Collection(JSON.parse(
         fs.readFileSync(path.resolve(__dirname, collectionsPath, './sample_collection.json').toString()))),
       body = collection.items.members[7].request.body,
       indentation = '  ',
@@ -62,7 +62,7 @@ describe('parseBody function', function () {
   });
 
   it('should return form-data params', function () {
-    const collection = new sdk.Collection(JSON.parse(
+    const collection = new Collection(JSON.parse(
         fs.readFileSync(path.resolve(__dirname, collectionsPath, './sample_collection.json').toString()))),
       body = collection.items.members[4].request.body,
       indentation = '  ',
@@ -99,7 +99,7 @@ describe('parseBody function', function () {
   });
 
   it('should return form-data params with a file', function () {
-    const collection = new sdk.Collection(JSON.parse(
+    const collection = new Collection(JSON.parse(
         fs.readFileSync(path.resolve(__dirname, collectionsPath, './sample_collection.json').toString()))),
       body = collection.items.members[26].request.body,
       indentation = '  ',
@@ -120,7 +120,7 @@ describe('parseBody function', function () {
   });
 
   it('should return binary data params with a file', function () {
-    const collection = new sdk.Collection(JSON.parse(
+    const collection = new Collection(JSON.parse(
         fs.readFileSync(path.resolve(__dirname, collectionsPath, './sample_collection.json').toString()))),
       body = collection.items.members[25].request.body,
       indentation = '  ',
@@ -132,7 +132,7 @@ describe('parseBody function', function () {
   });
 
   it('should return graphql params', function () {
-    const collection = new sdk.Collection(JSON.parse(
+    const collection = new Collection(JSON.parse(
         fs.readFileSync(path.resolve(__dirname, collectionsPath, './sample_collection.json').toString()))),
       body = collection.items.members[27].request.body,
       indentation = '  ',
@@ -147,7 +147,7 @@ describe('parseBody function', function () {
   });
 
   it('should return empty form data', function () {
-    const collection = new sdk.Collection(JSON.parse(
+    const collection = new Collection(JSON.parse(
         fs.readFileSync(path.resolve(__dirname, collectionsPath, './sample_collection.json').toString()))),
       body = collection.items.members[28].request.body,
       indentation = '  ',