+++ codegens/powershell-restmethod/test/unit/convert.test.js
@@ -1,5 +1,5 @@
 var expect = require('chai').expect,
-  sdk = require('postman-collection'),
+  { Request } = require('postman-collection/lib/collection/request'),
   exec = require('shelljs').exec,
   newman = require('newman'),
   parallel = require('async').parallel,
@@ -116,7 +116,7 @@ describe('Powershell-restmethod converter', function () {
     mainCollection.item.forEach(function (item) {
       // Skipping tests for Travis CI, till powershell dependency issue is sorted on travis
       it.skip(item.name, function (done) {
-        var request = new sdk.Request(item.request),
+        var request = new Request(item.request),
           collection = {
             item: [
               {
@@ -168,7 +168,7 @@ describe('Powershell-restmethod converter', function () {
           },
           'description': 'Description'
         },
-        pmRequest = new sdk.Request(request),
+        pmRequest = new Request(request),
         options = {
           requestTimeout: 10000,
           multiLine: true,
@@ -198,7 +198,7 @@ describe('Powershell-restmethod converter', function () {
     var request,
       options;
     it('should add a TimeoutSec argument when timeout is set to non zero value', function () {
-      request = new sdk.Request(mainCollection.item[0].request);
+      request = new Request(mainCollection.item[0].request);
       options = {
         requestTimeout: 1000
       };
@@ -213,7 +213,7 @@ describe('Powershell-restmethod converter', function () {
     });
 
     it('should not add a TimeoutSec argument when timeout is set to 0', function () {
-      request = new sdk.Request(mainCollection.item[0].request);
+      request = new Request(mainCollection.item[0].request);
       options = {
         requestTimeout: 0
       };
@@ -228,7 +228,7 @@ describe('Powershell-restmethod converter', function () {
     });
 
     it('should add a MaximumRedirection set to 0 argument when followRedirect is not allowed', function () {
-      request = new sdk.Request(mainCollection.item[0].request);
+      request = new Request(mainCollection.item[0].request);
       options = {
         followRedirect: false
       };
@@ -243,7 +243,7 @@ describe('Powershell-restmethod converter', function () {
     });
 
     it('should not add a MaximumRedirection argument when followRedirect is allowed', function () {
-      request = new sdk.Request(mainCollection.item[0].request);
+      request = new Request(mainCollection.item[0].request);
       options = {
         followRedirect: true
       };
@@ -258,7 +258,7 @@ describe('Powershell-restmethod converter', function () {
     });
 
     it('should default to mode raw when body mode is some random value', function () {
-      request = new sdk.Request(mainCollection.item[2].request);
+      request = new Request(mainCollection.item[2].request);
       request.body.mode = 'random';
       request.body[request.body.mode] = {};
       options = {};
@@ -272,7 +272,7 @@ describe('Powershell-restmethod converter', function () {
     });
 
     it('should generate snippet for file body mode', function () {
-      request = new sdk.Request({
+      request = new Request({
         'url': 'https://echo.getpostman.com/post',
         'method': 'POST',
         'body': {
@@ -297,7 +297,7 @@ describe('Powershell-restmethod converter', function () {
     });
 
     it('should trim header keys and not trim header values', function () {
-      var request = new sdk.Request({
+      var request = new Request({
         'method': 'GET',
         'header': [
           {
@@ -324,7 +324,7 @@ describe('Powershell-restmethod converter', function () {
     });
 
     it('should include graphql body in the snippet', function () {
-      var request = new sdk.Request({
+      var request = new Request({
         'method': 'POST',
         'header': [],
         'body': {
@@ -357,7 +357,7 @@ describe('Powershell-restmethod converter', function () {
     });
 
     it('should generate snippets(not error out) for requests with multiple/no file in formdata', function () {
-      var request = new sdk.Request({
+      var request = new Request({
         'method': 'POST',
         'header': [],
         'body': {
@@ -426,7 +426,7 @@ describe('Powershell-restmethod converter', function () {
     });
 
     it('should add content type if formdata field contains a content-type', function () {
-      var request = new sdk.Request({
+      var request = new Request({
         'method': 'POST',
         'body': {
           'mode': 'formdata',
@@ -462,7 +462,7 @@ describe('Powershell-restmethod converter', function () {
     });
 
     it('should generate valid snippet for single/double quotes in url', function () {
-      var request = new sdk.Request({
+      var request = new Request({
         'method': 'GET',
         'header': [],
         'url': {
@@ -499,7 +499,7 @@ describe('Powershell-restmethod converter', function () {
     });
 
     it('should generate valid snippet when single quotes in custom request method', function () {
-      var request = new sdk.Request({
+      var request = new Request({
         // eslint-disable-next-line quotes
         'method': "TEST';DIR;#'",
         'header': [],
@@ -538,7 +538,7 @@ describe('Powershell-restmethod converter', function () {
 
 
     it('should generate snippet for form data params with no type key present', function () {
-      var request = new sdk.Request({
+      var request = new Request({
         method: 'POST',
         header: [],
         url: {