+++ CHANGELOG.md
@@ -2,9 +2,24 @@
 
 ## [Unreleased]
 
-## [v1.10.1] - 2024-05-06
+## [v1.12.0] - 2024-07-22
 
-## [v1.10.0] - 2024-05-02
+### Chore
+
+-  Updated postman-collection sdk to version 4.4.0 in missing codegens.
+
+### Fixed
+
+-  Fix typo in Content-Header for audio/midi files in codegens.
+-  Added support for NTLM auth support in cURL codegen.
+
+## [v1.11.0] - 2024-07-10
+
+### Chore
+
+-   Updated postman-collection to v4.4.0.
+
+## [v1.10.1] - 2024-05-06
 
 ### Fixed
 
@@ -155,11 +170,13 @@ v1.0.0 (May 29, 2020)
 -   Add ES6 syntax support for NodeJS Request, NodeJS Native and NodeJS Unirest
 -   Fix snippet generation for powershell and jquery, where form data params had no type field
 
-[Unreleased]: https://github.com/postmanlabs/postman-code-generators/compare/v1.10.1...HEAD
+[Unreleased]: https://github.com/postmanlabs/postman-code-generators/compare/v1.12.0...HEAD
 
-[v1.10.1]: https://github.com/postmanlabs/postman-code-generators/compare/v1.10.0...v1.10.1
+[v1.12.0]: https://github.com/postmanlabs/postman-code-generators/compare/v1.11.0...v1.12.0
 
-[v1.10.0]: https://github.com/postmanlabs/postman-code-generators/compare/v1.9.0...v1.10.0
+[v1.11.0]: https://github.com/postmanlabs/postman-code-generators/compare/v1.10.1...v1.11.0
+
+[v1.10.1]: https://github.com/postmanlabs/postman-code-generators/compare/v1.10.0...v1.9.0
 
 [v1.9.0]: https://github.com/postmanlabs/postman-code-generators/compare/v1.8.0...v1.9.0